finished task1, task2 to go

+++ Project_3/Heap.h
@@ -5,7 +5,7 @@
 
 typedef struct heap_content{
 	int data; // Item d
-	int prior; // Item priority
+	float prior; // Item priority
 }HContent;
 
 typedef struct heap{
